requirements.txt
APScheduler==3.6.3
certifi==2020.12.5
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.20.1
Pillow==8.1.2
pycairo==1.20.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-telegram-bot==13.3
pytz==2021.1
PyYAML==6.0
six==1.15.0
squarify==0.4.3
tornado==6.1
tzlocal==2.1
